Move sheet importer into its own file
